projects/youtube_vedio_summerization/requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
google
youtube_transcript_api
reportlab
fpdf
